requirements.txt
openai-whisper
gTTS~=2.5.4
websockets~=15.0
playsound~=1.3.0
PyAudio~=0.2.14